requirements.txt
tornado==5.0.2
python-engineio
python_socketio
six==1.10.0
uuid~=1.30
vosk~=0.3.32
paho-mqtt~=1.6.1
websockets~=10.0